% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp_insert.R
\name{bcp_insert}
\alias{bcp_insert}
\title{Bulk Insert Data into SQL Server}
\usage{
bcp_insert(
  input_data,
  server,
  database,
  table,
  schema = "dbo",
  overwrite = FALSE,
  username = NULL,
  password = NULL,
  trusted_connection = TRUE,
  bcp_path = "bcp",
  batch_size = 1000000L,
  tmpdir = tempdir(),
  packet_size = 4096L
)
}
\arguments{
\item{input_data}{A data.frame, data.table, or path to a .parquet file.}

\item{server}{SQL Server instance name.}

\item{database}{Database name.}

\item{table}{Target table name.}

\item{schema}{Target schema name (default "dbo").}

\item{overwrite}{Boolean. If TRUE, table is TRUNCATED before insert. Default FALSE.}

\item{username}{SQL authentication username (optional).}

\item{password}{SQL authentication password (optional).}

\item{trusted_connection}{Boolean. Use Windows Authentication? Default TRUE.}

\item{bcp_path}{Path to bcp executable. Default assumes it is in system PATH.}

\item{batch_size}{Number of rows per batch.}

\item{tmpdir}{Directory for temporary files.}

\item{packet_size}{Network packet size in bytes.}
}
\description{
Bulk Insert Data into SQL Server
}
