% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_for_bcp}
\alias{format_for_bcp}
\title{Sanitize Data for BCP}
\usage{
format_for_bcp(dt)
}
\description{
Only handles Logical -> Integer conversion.
Date/Time/NA formatting is now handled by fwrite for performance.
}
\keyword{internal}
